requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
isort==5.5.3
lazy-object-proxy==1.5.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
pylint==2.6.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.53.0
django-ckeditor==6.0.0
mysqlclient==2.0.1